+++ src/csidh_params.rs
@@ -141,6 +141,54 @@ impl CsidhParams<LIMBS_1792, 201, PrimeCsidh1792> {
 }
 
 impl<const LIMBS: usize, const N: usize, MOD: ConstMontyParams<LIMBS>> CsidhParams<LIMBS, N, MOD> {
+    /// Constructs custom parameters. The caller is responsible for the validity of the
+    /// parameters. Valid parameters respect the following rules:
+    ///
+    /// - `lis` must be an array of mutually different prime numbers and contain the number 3.
+    /// Their product, multiplied by 4, minus 1, must be a prime number that is called p.
+    /// - `p_minus_1_over_2` must be equal to (p-1)/2.
+    /// - `inverse_of_4` must be the inverse of 4 in the field of cardinality p.
+    /// - The LIMBS generic given to `p_minus_1_over_2` and `inverse_of_4` must be the same, and
+    /// must be big enough to store numbers up to p. It is advised to use the smallest LIMBS number
+    /// that satisfies this condition to minimize execution time. This translates to the following:
+    ///     - LIMBS = min([0, usize::MAX]) such that 2<sup>(LIMBS * target_pointer_width)</sup> > p
+    ///
+    /// It is **unsound** to use invalid parameters. No validation is performed by the callee.
+    /// **Use with care.**
+    ///
+    /// # Example
+    ///
+    /// To construct the parameters from the prime numbers [3, 5, 7]:
+    ///
+    /// ```
+    /// use csidh::{
+    ///     impl_modulus, ConstMontyForm, CsidhParams, PrivateKey, PublicKey, SharedSecret, Uint
+    /// };
+    ///
+    /// # fn main() {
+    /// const LIMBS_3_5_7: usize = 1;
+    /// impl_modulus!(Prime419, Uint<LIMBS_3_5_7>, "00000000000001a3");
+    ///
+    /// let lis = [3, 5, 7];
+    /// let p_minus_1_over_2 = Uint::from(209u32);
+    /// let inverse_of_4: ConstMontyForm<Prime419, LIMBS_3_5_7> =
+    ///     ConstMontyForm::new(&Uint::from(105u32));
+    ///
+    /// let params = CsidhParams::new(lis, p_minus_1_over_2, inverse_of_4);
+    /// # }
+    /// ```
+    pub const fn new(
+        lis: [u64; N],
+        p_minus_1_over_2: Uint<LIMBS>,
+        inverse_of_4: ConstMontyForm<MOD, LIMBS>,
+    ) -> Self {
+        CsidhParams {
+            lis,
+            p_minus_1_over_2,
+            inverse_of_4,
+        }
+    }
+
     pub(crate) const fn lis(self) -> [u64; N] {
         self.lis
     }