+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `PublicKey::from` and `SharedSecret::from` now take about 20% less time.
+
 ## [0.3] - 2024-06-02
 
 ### Changed