+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `CsidhParams::1024` and `CsidhParams::1792` are now available.
 The prime numbers used are not standard, as their is no such standard.
+- `CsidhParams::new` is now available, allowing to create custom params.
 
 ### Changed